implementing wire protocol

+++ shard/write.go
@@ -22,7 +22,7 @@ type (
 // shardBit is the number of bits used in the shard index
 // the number of shards is 1<<shardBit
 // the shard mask is 1<<shardBit - 1
-func NewWriter(dir string, shardBit uint, shardFunc HashFunc) (*Writer, error) {
+func NewWriter(dir string, shardBit uint8, shardFunc HashFunc) (*Writer, error) {
 	if shardBit > 10 {
 		return nil, errors.New("shardBit should be no more than 10")
 	}
